add spec for create_new_device_alert, feature checking for existing device tests

+++ spec/services/user_alerts/alert_user_about_new_device_spec.rb
@@ -6,24 +6,32 @@ RSpec.describe UserAlerts::AlertUserAboutNewDevice do
     let(:disavowal_token) { 'the_disavowal_token' }
     let(:device) { create(:device, user: user) }
 
-    it 'sends an email to all confirmed email addresses' do
-      user.email_addresses.destroy_all
-      confirmed_email_addresses = create_list(:email_address, 2, user: user)
-      create(:email_address, user: user, confirmed_at: nil)
+    if FeatureManagement.aggregate_new_device_alerts?
+      it 'sets the user sign_in_new_device value to current datetime' do
+        described_class.call(user, device, disavowal_token)
+        expect(user.sign_in_new_device).not_to be_nil
+      end
+    else
 
-      described_class.call(user, device, disavowal_token)
+      it 'sends an email to all confirmed email addresses' do
+        user.email_addresses.destroy_all
+        confirmed_email_addresses = create_list(:email_address, 2, user: user)
+        create(:email_address, user: user, confirmed_at: nil)
 
-      expect_delivered_email_count(2)
-      expect_delivered_email(
-        to: [confirmed_email_addresses[0].email],
-        subject: t('user_mailer.new_device_sign_in.subject', app_name: APP_NAME),
-        body: [disavowal_token],
-      )
-      expect_delivered_email(
-        to: [confirmed_email_addresses[1].email],
-        subject: t('user_mailer.new_device_sign_in.subject', app_name: APP_NAME),
-        body: [disavowal_token],
-      )
+        described_class.call(user, device, disavowal_token)
+
+        expect_delivered_email_count(2)
+        expect_delivered_email(
+          to: [confirmed_email_addresses[0].email],
+          subject: t('user_mailer.new_device_sign_in.subject', app_name: APP_NAME),
+          body: [disavowal_token],
+        )
+        expect_delivered_email(
+          to: [confirmed_email_addresses[1].email],
+          subject: t('user_mailer.new_device_sign_in.subject', app_name: APP_NAME),
+          body: [disavowal_token],
+        )
+      end
     end
   end
 end
